refactor: rename default_mode leaf files to their responsibilities
shortcuts/apply.rs (the ShortcutSet::Apply appliers), mouse/webview/router.rs
(the stateful pointer router over the stateless CEF helpers), ui/
shell_surface.rs (shell-surface spawn under UiRoot); the 64-line paste
applier merges into paste.rs. Identifier renames follow in phase 2.
