+++ src/input/shortcuts/apply.rs
